+++ src/extension.ts
@@ -3,27 +3,27 @@
 import * as vscode from 'vscode';
 
 import { EnvironmentsProvider } from './environments';
-import { GitHelper } from './git';
-import { serverSettings } from './interfaces';
-import { PackageManager } from './managers/packagemanager';
-import { Constants, getDirectoryName, showErrorMessage, showInformationMessage } from './constants';
+import { PackageExplorer, PackageProvider } from './packageExplorer';
+import { packageSettings, serverSettings } from './interfaces';
+import { Constants, getDirectoryName, hash, showErrorMessage } from './constants';
 import { ClioManager } from './managers/cliomanager';
 
 let terminalLog: vscode.OutputChannel;
 let terminal: vscode.Terminal;
-let environments: serverSettings[] | undefined;
+const bpmPackagesPattern = `${hash}_bpmPackages`;
 
 function checkWorkspaceSettings() {
 	const serverConfig = vscode.workspace.getConfiguration('cwSettings');
-	environments = serverConfig.get<serverSettings[]>('cwEnvironments');
+	Constants.environments = serverConfig.get<serverSettings[]>('cwEnvironments');
 
-	if (environments && environments.length > 0) {
+	if (Constants.environments && Constants.environments.length > 0) {
 		vscode.commands.executeCommand('setContext', 'isShowContextMenu', true);
 	} else {
 		vscode.commands.executeCommand('setContext', 'isShowContextMenu', false);
 	}
 
 	vscode.commands.executeCommand('bpmsoftEnvironments.refreshEntry');
+	vscode.commands.executeCommand('packagesExplorer.refreshEntry');
 }
 
 // This method is called when your extension is activated
@@ -41,20 +41,20 @@ export function activate(context: vscode.ExtensionContext) {
 
 	Constants.extensionPath = context.extensionPath;
 
-	const gitHelper = new GitHelper(terminalLog);
-	const pm = new PackageManager(terminalLog);
+	const pe = new PackageExplorer(terminalLog);
 	const cm = new ClioManager(terminalLog);
 
-	pm.onCommandExecuteError = (message: string, showbutton: boolean) => showErrorMessage(message, showbutton);
-	pm.onCommandExecuteComplete = (message: string, showbutton: boolean) => showInformationMessage(message, showbutton);
-	cm.onCommandExecuteError = (message: string, showbutton: boolean) => showErrorMessage(message, showbutton);
+	cm.onCommandExecuteError = (message: string, showbutton: boolean, executeLogFilePath: string | undefined) =>
+		showErrorMessage(message, showbutton, executeLogFilePath);
 
 	const environmentsProvider = new EnvironmentsProvider();
+	const packageProvider = new PackageProvider();
 	vscode.window.registerTreeDataProvider('bpmsoftEnvironments', environmentsProvider);
+	vscode.window.registerTreeDataProvider('packagesExplorer', packageProvider);
 
 	vscode.commands.registerCommand('bpmsoftEnvironments.refreshEntry', () => {
-		if (environments) {
-			environments?.forEach(env => {
+		if (Constants.environments) {
+			Constants.environments?.forEach(env => {
 				var isReg = context.globalState.get(env.id);
 				if (isReg) {
 					env.isRegister = true;
@@ -62,11 +62,37 @@ export function activate(context: vscode.ExtensionContext) {
 					env.isRegister = false;
 				}
 			});
-			environmentsProvider.refresh(environments);
+			environmentsProvider.refresh(Constants.environments);
 		} else {
 			environmentsProvider.refresh([]);
 		}
 	});
+
+	vscode.commands.registerCommand('packagesExplorer.refreshEntry', () => {
+		let packages = context.globalState.get<Array<packageSettings>>(bpmPackagesPattern) ?? new Array();
+		if (packages) {
+			packageProvider.refresh(packages);
+		} else {
+			packageProvider.refresh([]);
+		}
+	});
+
+	context.subscriptions.push(vscode.commands.registerCommand('packagesExplorer.remove', (pkg: packageSettings) => {
+		let packages = context.globalState.get<Array<packageSettings>>(bpmPackagesPattern) ?? new Array();
+		if (packages.length > 0) {
+			packages = packages.filter(item => { return item.folderName !== pkg.folderName; });
+			context.globalState.update(bpmPackagesPattern, packages);
+			vscode.commands.executeCommand('packagesExplorer.refreshEntry');
+		}
+	}));
+
+	context.subscriptions.push(vscode.commands.registerCommand('packagesExplorer.create', (pkg: packageSettings) => {
+		pe.create(pkg.targetFolderPath);
+	}));
+
+	context.subscriptions.push(vscode.commands.registerCommand('packagesExplorer.createandsend', (pkg: packageSettings) => {
+		pe.createAndSend(pkg.targetFolderPath);
+	}));
 
 	context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(event => {
 		checkWorkspaceSettings();
@@ -75,7 +101,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(vscode.commands.registerCommand(
 		'clio.openLastLog',
 		() => {
-			const folderUri = vscode.Uri.file(Constants.extensionPath + Constants.executeLogFileName);
+			const folderUri = vscode.Uri.file(cm.getLastExecuteLogPath());
 			vscode.commands.executeCommand(`vscode.openFolder`, folderUri);
 		}));
 
@@ -135,81 +161,26 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.executeCommand('bpmsoftEnvironments.refreshEntry');
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('cliowrapper.package.create', (uri: vscode.Uri) => {
-		const config = vscode.workspace.getConfiguration('clio');
-		const outputPath = config.get<string>('outputPath');
-
-		if (environments) {
-			const branches = environments.map(({ gitBranchName }) => gitBranchName ?? '');
-			vscode.window.withProgress(
-				{
-					location: vscode.ProgressLocation.Notification,
-					title: `Creating package: ${getDirectoryName(uri.fsPath)}.gz`
-				},
-				async () => {
-					vscode.commands.executeCommand('setContext', 'isShowContextMenu', false);
-					if (await gitHelper.isPermittedBranch(branches)) {
-						if (await pm.createPackage(uri.fsPath)) {
-							if (outputPath) {
-								vscode.env.openExternal(vscode.Uri.file(outputPath));
-							}
-						}
-					}
-					vscode.commands.executeCommand('setContext', 'isShowContextMenu', true);
-				}
-			);
-		} else {
-			vscode.window.showInformationMessage(
-				"Command execute failed: check you .code-workspace file."
-			);
+	context.subscriptions.push(vscode.commands.registerCommand('cliowrapper.package.add.packagesExplorer', (uri: vscode.Uri) => {
+		let packages = context.globalState.get<Array<packageSettings>>(bpmPackagesPattern) ?? new Array();
+		const newPackage: packageSettings = {
+			folderName: getDirectoryName(uri.fsPath),
+			targetFolderPath: uri.fsPath,
+			targetEnviroment: undefined
+		};
+		if (!packages.find(p => p.targetFolderPath === uri.fsPath)) {
+			packages.push(newPackage);
+			context.globalState.update(bpmPackagesPattern, packages);
+			vscode.commands.executeCommand('packagesExplorer.refreshEntry');
 		}
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('cliowrapper.package.createandsend', (uri: vscode.Uri) => {
-		const currentBranch = gitHelper.getCurrentBranch();
-		if (environments && currentBranch !== '') {
-			const arr = environments.filter(e => e.gitBranchName === currentBranch && e.isEnable && e.isRegister)?.map(({ id }) => id);
-			const quickPickItems = arr.map(item => ({ label: item, iconPath: new vscode.ThemeIcon('device-desktop') }));
-			const qp = vscode.window.createQuickPick();
-			qp.canSelectMany = false;
-			qp.items = quickPickItems;
-			qp.onDidChangeSelection(async selection => {
-				const serverId = selection[0].label;
-				const serverConfig = environments?.find(e => e.id === serverId);
-				qp.hide();
+	context.subscriptions.push(vscode.commands.registerCommand('cliowrapper.package.create', (uri: vscode.Uri) => {
+		pe.create(uri.fsPath);
+	}));
 
-				if (serverConfig && serverConfig.gitBranchName) {
-					vscode.window.withProgress(
-						{
-							location: vscode.ProgressLocation.Notification,
-							title: 'Current operation'
-						},
-						async (progress) => {
-							vscode.commands.executeCommand('setContext', 'isShowContextMenu', false);
-							if (await gitHelper.checkBranch(serverConfig.gitBranchName!)) {
-								progress.report({
-									message: `creating package ${getDirectoryName(uri.fsPath)}.gz`
-								});
-								if (await pm.createPackage(uri.fsPath)) {
-									progress.report({
-										message: `sending package ${getDirectoryName(uri.fsPath)}.gz`,
-										increment: 50
-									});
-									await pm.pushPackage({ targetFolderPath: uri.fsPath, targetEnviroment: serverConfig.id });
-								}
-							}
-							vscode.commands.executeCommand('setContext', 'isShowContextMenu', true);
-						}
-					);
-				} else {
-					vscode.window.showInformationMessage(
-						"Command execute failed: check you .code-workspace file."
-					);
-				}
-			});
-			qp.onDidHide(() => qp.dispose());
-			qp.show();
-		}
+	context.subscriptions.push(vscode.commands.registerCommand('cliowrapper.package.createandsend', (uri: vscode.Uri) => {
+		pe.createAndSend(uri.fsPath);
 	}));
 }