+++ src/managers/cliomanager.ts
@@ -1,20 +1,24 @@
 import * as vscode from 'vscode';
 import { TerminalWrapper } from '../terminal/terminalwrapper';
 import { serverSettings } from '../interfaces';
-import { isNullOrWhitespace } from '../constants';
+import { Constants, isNullOrWhitespace } from '../constants';
 
 export class ClioManager {
     private terminal: TerminalWrapper;
 
-    public onCommandExecuteError?: (message: string, showbutton: boolean) => void;
-    public onCommandExecuteComplete?: (message: string, showbutton: boolean) => void;
+    public onCommandExecuteError?: (message: string, showbutton: boolean, outputPathLog: string | undefined) => void;
+    public onCommandExecuteComplete?: (message: string, showbutton: boolean, outputPathLog: string | undefined) => void;
 
     constructor(private terminalLog: vscode.OutputChannel) {
-        this.terminal = new TerminalWrapper();
+        this.terminal = new TerminalWrapper(Constants.extensionPath);
+    }
+
+    public getLastExecuteLogPath(): string {
+        return this.terminal.executeLogFilePath;
     }
 
     public OpenSettings() {
-        this.terminal.callInInteractiveTerminalWithoutLog('clio open-settings');
+        this.terminal.executeCommandWithoutLog('clio open-settings');
     }
 
     public async WebAppRegister(server: serverSettings): Promise<boolean> {
@@ -29,7 +33,7 @@ export class ClioManager {
                 password: true
             });
             if (!isNullOrWhitespace(passwordQuery)) {
-                return await this.terminal.callInInteractiveTerminal(
+                return await this.terminal.executeCommand(
                     `clio reg-web-app ${server.id} -u ${server.url} -l ${loginQuery} -p ${passwordQuery}`
                 );
             }
@@ -39,19 +43,19 @@ export class ClioManager {
 
     public async WebAppUnregister(server: serverSettings, isLogEnabled: boolean) {
         if(isLogEnabled){
-            const result = await this.terminal.callInInteractiveTerminal(`clio unreg-web-app ${server.id}`);
+            const result = await this.terminal.executeCommand(`clio unreg-web-app ${server.id}`);
             if (!result && this.onCommandExecuteError) {
-                this.onCommandExecuteError('Unregister web app failed.', true);
+                this.onCommandExecuteError('Unregister web app failed.', true, this.terminal.executeLogFilePath);
             }
         }else{
-            await this.terminal.callInInteractiveTerminalWithoutLog(`clio unreg-web-app ${server.id}`);
+            await this.terminal.executeCommandWithoutLog(`clio unreg-web-app ${server.id}`);
         }
     }
 
     public async WebAppPing(server: serverSettings): Promise<boolean> {
-        const result = await this.terminal.callInInteractiveTerminal(`clio ping ${server.id}`);
+        const result = await this.terminal.executeCommand(`clio ping ${server.id}`);
         if (!result && this.onCommandExecuteError) {
-            this.onCommandExecuteError('Ping web app failed.', true);
+            this.onCommandExecuteError('Ping web app failed.', true, this.terminal.executeLogFilePath);
         }
         return result;
     }
@@ -62,9 +66,9 @@ export class ClioManager {
                 `You cannot execute this command because server ${server.id} is disabled.`
             );
         } else {
-            const result = await this.terminal.callInInteractiveTerminal(`clio restart-web-app ${server.id}`);
+            const result = await this.terminal.executeCommand(`clio restart-web-app ${server.id}`);
             if (!result && this.onCommandExecuteError) {
-                this.onCommandExecuteError('Restart web app failed.', true);
+                this.onCommandExecuteError('Restart web app failed.', true, this.terminal.executeLogFilePath);
             }
         }
     }
@@ -75,9 +79,9 @@ export class ClioManager {
                 "You cannot execute this command because server " + server.id + " is disabled."
             );
         } else {
-            const result = await this.terminal.callInInteractiveTerminal(`clio clear-redis-db ${server.id}`);
+            const result = await this.terminal.executeCommand(`clio clear-redis-db ${server.id}`);
             if (!result && this.onCommandExecuteError) {
-                this.onCommandExecuteError('Clear redis db failed.', true);
+                this.onCommandExecuteError('Clear redis db failed.', true, this.terminal.executeLogFilePath);
             }
         }
     }
@@ -88,9 +92,9 @@ export class ClioManager {
                 "You cannot execute this command because server " + server.id + " is disabled."
             );
         } else {
-            const result = await this.terminal.callInInteractiveTerminal(`clio compile-configuration ${server.id}`);
+            const result = await this.terminal.executeCommand(`clio compile-configuration ${server.id}`);
             if (!result && this.onCommandExecuteError) {
-                this.onCommandExecuteError('Compile configuration failed.', true);
+                this.onCommandExecuteError('Compile configuration failed.', true, this.terminal.executeLogFilePath);
             }
         }
     }